refactor(components): extract tooltip logic into custom hook
